Refactor Language settings into connected container and stateless functional component Tidy up other settings refactors.
